fix(#1054): handle wipe out command properly
